removed unneccessary casts

+++ 1.9.4/src/main/java/net/geforcemods/securitycraft/blocks/BlockFakeLava.java
@@ -60,7 +60,7 @@ public class BlockFakeLava extends BlockDynamicLiquid implements IIntersectable 
 
     public void updateTick(World worldIn, BlockPos pos, IBlockState state, Random rand)
     {
-        int i = ((Integer)state.getValue(LEVEL)).intValue();
+        int i = state.getValue(LEVEL).intValue();
         byte b0 = 1;
 
         if (this.blockMaterial == Material.LAVA && !worldIn.provider.doesWaterVaporize())
@@ -111,7 +111,7 @@ public class BlockFakeLava extends BlockDynamicLiquid implements IIntersectable 
                 {
                     l = 0;
                 }
-                else if (iblockstate2.getBlock().getMaterial(iblockstate2) == this.blockMaterial && ((Integer)iblockstate2.getValue(LEVEL)).intValue() == 0)
+                else if (iblockstate2.getBlock().getMaterial(iblockstate2) == this.blockMaterial && iblockstate2.getValue(LEVEL).intValue() == 0)
                 {
                     l = 0;
                 }
@@ -227,7 +227,7 @@ public class BlockFakeLava extends BlockDynamicLiquid implements IIntersectable 
                 BlockPos blockpos1 = pos.offset(enumfacing1);
                 IBlockState iblockstate = worldIn.getBlockState(blockpos1);
 
-                if (!this.isBlocked(worldIn, blockpos1, iblockstate) && (iblockstate.getBlock().getMaterial(iblockstate) != this.blockMaterial || ((Integer)iblockstate.getValue(LEVEL)).intValue() > 0))
+                if (!this.isBlocked(worldIn, blockpos1, iblockstate) && (iblockstate.getBlock().getMaterial(iblockstate) != this.blockMaterial || iblockstate.getValue(LEVEL).intValue() > 0))
                 {
                     if (!this.isBlocked(worldIn, blockpos1.down(), iblockstate))
                     {
@@ -262,7 +262,7 @@ public class BlockFakeLava extends BlockDynamicLiquid implements IIntersectable 
             BlockPos blockpos1 = pos.offset(enumfacing);
             IBlockState iblockstate = worldIn.getBlockState(blockpos1);
 
-            if (!this.isBlocked(worldIn, blockpos1, iblockstate) && (iblockstate.getBlock().getMaterial(iblockstate) != this.blockMaterial || ((Integer)iblockstate.getValue(LEVEL)).intValue() > 0))
+            if (!this.isBlocked(worldIn, blockpos1, iblockstate) && (iblockstate.getBlock().getMaterial(iblockstate) != this.blockMaterial || iblockstate.getValue(LEVEL).intValue() > 0))
             {
                 int j;
