+++ 1.9.4/src/main/java/net/geforcemods/securitycraft/blocks/BlockFakeWater.java
@@ -55,7 +55,7 @@ public class BlockFakeWater extends BlockDynamicLiquid{
 
     public void updateTick(World worldIn, BlockPos pos, IBlockState state, Random rand)
     {
-        int i = ((Integer)state.getValue(LEVEL)).intValue();
+        int i = state.getValue(LEVEL).intValue();
         byte b0 = 1;
 
         if (this.blockMaterial == Material.LAVA && !worldIn.provider.doesWaterVaporize())
@@ -106,7 +106,7 @@ public class BlockFakeWater extends BlockDynamicLiquid{
                 {
                     l = 0;
                 }
-                else if (iblockstate2.getBlock().getMaterial(iblockstate2) == this.blockMaterial && ((Integer)iblockstate2.getValue(LEVEL)).intValue() == 0)
+                else if (iblockstate2.getBlock().getMaterial(iblockstate2) == this.blockMaterial && iblockstate2.getValue(LEVEL).intValue() == 0)
                 {
                     l = 0;
                 }
@@ -222,7 +222,7 @@ public class BlockFakeWater extends BlockDynamicLiquid{
                 BlockPos blockpos1 = pos.offset(enumfacing1);
                 IBlockState iblockstate = worldIn.getBlockState(blockpos1);
 
-                if (!this.isBlocked(worldIn, blockpos1, iblockstate) && (iblockstate.getBlock().getMaterial(iblockstate) != this.blockMaterial || ((Integer)iblockstate.getValue(LEVEL)).intValue() > 0))
+                if (!this.isBlocked(worldIn, blockpos1, iblockstate) && (iblockstate.getBlock().getMaterial(iblockstate) != this.blockMaterial || iblockstate.getValue(LEVEL).intValue() > 0))
                 {
                     if (!this.isBlocked(worldIn, blockpos1.down(), iblockstate))
                     {
@@ -257,7 +257,7 @@ public class BlockFakeWater extends BlockDynamicLiquid{
             BlockPos blockpos1 = pos.offset(enumfacing);
             IBlockState iblockstate = worldIn.getBlockState(blockpos1);
 
-            if (!this.isBlocked(worldIn, blockpos1, iblockstate) && (iblockstate.getBlock().getMaterial(iblockstate) != this.blockMaterial || ((Integer)iblockstate.getValue(LEVEL)).intValue() > 0))
+            if (!this.isBlocked(worldIn, blockpos1, iblockstate) && (iblockstate.getBlock().getMaterial(iblockstate) != this.blockMaterial || iblockstate.getValue(LEVEL).intValue() > 0))
             {
                 int j;
